packages/python-sdk/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "z402-sdk"
version = "0.1.0"
description = "Python SDK for Z402 - Optimized for AI agents and automation"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Z402 Team", email = "team@z402.io"}
]
keywords = ["zcash", "payment", "cryptocurrency", "z402", "ai-agent", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "aiohttp>=3.9.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]
langchain = [
    "langchain>=0.1.0",
]
autogpt = [
    "autogpt>=0.1.0",
]
fastapi = [
    "fastapi>=0.104.0",
]
flask = [
    "flask>=3.0.0",
]

[project.scripts]
z402 = "z402.cli:app"

[project.urls]
Homepage = "https://z402.io"
Documentation = "https://docs.z402.io/python"
Repository = "https://github.com/z402/python-sdk"
"Bug Tracker" = "https://github.com/z402/python-sdk/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["z402*"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "I", "N", "W"]
